+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Feat: Allow setting SDK info (name & version) in manifest (#2016)
 
 ## 6.0.0-beta.3